Fix #870: Move classes around to avoid SAIL dependency on RepositoryManager
