+++ core/queryalgebra/model/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/federation/AbstractFederatedServiceResolver.java
@@ -11,7 +11,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.eclipse.rdf4j.query.QueryEvaluationException;
-import org.eclipse.rdf4j.repository.RepositoryException;
 
 /**
  * Base class for {@link FederatedServiceResolver} which takes care for lifecycle management of produced
