+++ core/repository/api/src/main/java/org/eclipse/rdf4j/repository/RepositoryResolverClient.java
@@ -5,14 +5,15 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/org/documents/edl-v10.php.
  *******************************************************************************/
-package org.eclipse.rdf4j.repository.sail.config;
+package org.eclipse.rdf4j.repository;
 
 /**
  * Interface used by factory classes that need access to other repositories by their id's.
  * 
  * @author Dale Visser
- * @deprecated since 2.3 use {@link org.eclipse.rdf4j.repository.RepositoryResolverClient}
+ * @since 2.3
  */
-@Deprecated
-public interface RepositoryResolverClient extends org.eclipse.rdf4j.repository.RepositoryResolverClient {
+public interface RepositoryResolverClient {
+
+	void setRepositoryResolver(RepositoryResolver resolver);
 }
