+++ core/repository/sail/src/main/java/org/eclipse/rdf4j/repository/sail/config/RepositoryResolver.java
@@ -15,8 +15,10 @@ import org.eclipse.rdf4j.repository.config.RepositoryConfigException;
  * Gets local repositories using a simple identifier string.
  * 
  * @author Dale Visser
+ * @deprecated since 2.3 use {@link org.eclipse.rdf4j.repository.RepositoryResolver}
  */
-public interface RepositoryResolver {
+@Deprecated
+public interface RepositoryResolver extends org.eclipse.rdf4j.repository.RepositoryResolver {
 
 	Repository getRepository(String memberID)
 		throws RepositoryException, RepositoryConfigException;
