+++ core/sail/federation/src/main/java/org/eclipse/rdf4j/sail/federation/AbstractFederationConnection.java
@@ -55,8 +55,8 @@ import org.eclipse.rdf4j.repository.RepositoryException;
 import org.eclipse.rdf4j.repository.RepositoryResult;
 import org.eclipse.rdf4j.repository.filters.AccurateRepositoryBloomFilter;
 import org.eclipse.rdf4j.repository.filters.RepositoryBloomFilter;
-import org.eclipse.rdf4j.repository.sail.config.RepositoryResolver;
-import org.eclipse.rdf4j.repository.sail.config.RepositoryResolverClient;
+import org.eclipse.rdf4j.repository.RepositoryResolver;
+import org.eclipse.rdf4j.repository.RepositoryResolverClient;
 import org.eclipse.rdf4j.sail.SailConnection;
 import org.eclipse.rdf4j.sail.SailException;
 import org.eclipse.rdf4j.sail.federation.optimizers.EmptyPatternOptimizer;
