+++ core/sail/nativerdf/src/main/java/org/eclipse/rdf4j/sail/nativerdf/NativeStore.java
@@ -24,8 +24,8 @@ import org.eclipse.rdf4j.query.algebra.evaluation.EvaluationStrategy;
 import org.eclipse.rdf4j.query.algebra.evaluation.EvaluationStrategyFactory;
 import org.eclipse.rdf4j.query.algebra.evaluation.federation.FederatedServiceResolver;
 import org.eclipse.rdf4j.query.algebra.evaluation.federation.FederatedServiceResolverClient;
-import org.eclipse.rdf4j.query.algebra.evaluation.federation.FederatedServiceResolverImpl;
 import org.eclipse.rdf4j.query.algebra.evaluation.impl.StrictEvaluationStrategyFactory;
+import org.eclipse.rdf4j.repository.sparql.federation.SPARQLServiceResolver;
 import org.eclipse.rdf4j.sail.NotifyingSailConnection;
 import org.eclipse.rdf4j.sail.SailException;
 import org.eclipse.rdf4j.sail.base.SailSource;
@@ -81,7 +81,7 @@ public class NativeStore extends AbstractNotifyingSail implements FederatedServi
 	private FederatedServiceResolver serviceResolver;
 
 	/** dependent life cycle */
-	private FederatedServiceResolverImpl dependentServiceResolver;
+	private SPARQLServiceResolver dependentServiceResolver;
 
 	/**
 	 * Lock manager used to prevent concurrent {@link #getTransactionLock(IsolationLevel)} calls.
@@ -197,7 +197,7 @@ public class NativeStore extends AbstractNotifyingSail implements FederatedServi
 	public synchronized FederatedServiceResolver getFederatedServiceResolver() {
 		if (serviceResolver == null) {
 			if (dependentServiceResolver == null) {
-				dependentServiceResolver = new FederatedServiceResolverImpl();
+				dependentServiceResolver = new SPARQLServiceResolver();
 			}
 			setFederatedServiceResolver(dependentServiceResolver);
 		}
